rework apis
update go version
error passtrought
idea run configurations
request examples

+++ storage-service/src/service/ApiHandlerService.go
@@ -4,20 +4,19 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/adrg/strutil/metrics"
 	"log/slog"
-	"strings"
 	"time"
 
+	"github.com/adrg/strutil/metrics"
+	"github.com/weoses/memelo/common/commonhelper"
+
 	"github.com/adrg/strutil"
-	"github.com/gdexlab/go-render/render"
 	"github.com/go-playground/validator/v10"
 	"github.com/google/uuid"
-	"github.com/labstack/echo/v4"
-	"mine.local/ocr-gallery/apispec/meme-storage/server"
-	"mine.local/ocr-gallery/common/commonconst"
-	"mine.local/ocr-gallery/storage-service/entity"
-	"mine.local/ocr-gallery/storage-service/helper"
+	"github.com/weoses/memelo/apispec/meme-storage/server"
+	"github.com/weoses/memelo/common/commonconst"
+	"github.com/weoses/memelo/storage-service/entity"
+	"github.com/weoses/memelo/storage-service/helper"
 )
 
 type ApiHandler struct {
@@ -31,6 +30,10 @@ func (a *ApiHandler) DeleteMeme(ctx context.Context, request server.DeleteMemeRe
 	item, err := a.metaStorage.GetById(ctx, request.AccountId, request.MemeId)
 	if err != nil {
 		return nil, err
+	}
+
+	if item == nil {
+		return server.DeleteMeme200Response{}, nil
 	}
 
 	err = a.metaStorage.DeleteById(ctx, item.AccountId, item.ImageId)
@@ -51,52 +54,53 @@ func (a *ApiHandler) CreateMeme(
 	request server.CreateMemeRequestObject,
 ) (server.CreateMemeResponseObject, error) {
 	image := request.Body
+	accountId := request.AccountId
 
 	slog.Info("CreateMeme",
-		commonconst.ACCOUNTID_LOG, request.AccountId)
+		commonconst.ACCOUNTID_LOG, accountId)
 
 	idUuid, _ := uuid.NewRandom()
-	if len(*image.ImageBase64) == 0 {
+	if len(image.ImageBase64) == 0 {
 		return nil, errors.New("image is empty")
 	}
 
 	hash := helper.CalcHash(request.Body.ImageBase64)
-	hashDuplicate, err := a.findHashDuplicates(ctx, hash)
+	hashDuplicate, err := a.findHashDuplicates(ctx, accountId, hash)
 	if err != nil {
 		return nil, fmt.Errorf("failed to find hash duplicates : %w", err)
 	}
 
-	if hashDuplicate != nil {
-		return a.HandleDuplicate(ctx, server.DuplicateHash, hashDuplicate, request)
+	if hashDuplicate != nil && len(hashDuplicate) > 0 {
+		return a.HandleDuplicate(ctx, server.DuplicateHash, hashDuplicate[0], request)
 	}
 
-	reqImage := helper.ImageToEntity2(request.Body)
-	ocrResult, err := a.ocr.DoOcr(ctx, idUuid, reqImage)
+	reqImage := helper.ApiImageToEntity(request.Body)
+	ocrResult, err := a.ocr.DoOcr(ctx, reqImage)
 	if err != nil {
 		return nil, fmt.Errorf("failed to do ocr : %w", err)
 	}
 
 	ocrTextResult := ocrResult.OcrText
 	slog.Info("CreateMeme: ocr result",
-		commonconst.ACCOUNTID_LOG, request.AccountId,
+		commonconst.ACCOUNTID_LOG, accountId,
 		"id", idUuid,
 		"ocrText", ocrTextResult)
 
-	contentDuplicate, err := a.findContentDuplicates(ctx, ocrResult)
+	contentDuplicate, err := a.findContentDuplicates(ctx, accountId, ocrResult)
 	if err != nil {
 		return nil, err
 	}
 
-	if strings.TrimSpace(ocrTextResult) == "" {
-		return nil, errors.New("no text on image")
-	}
+	//if strings.TrimSpace(ocrTextResult) == "" {
+	//	return nil, errors.New("no text on image")
+	//}
 
 	if contentDuplicate != nil {
 		contentDuplicateTextResult := contentDuplicate.Result
 		similarity := strutil.Similarity(ocrTextResult, contentDuplicateTextResult, metrics.NewLevenshtein())
 
 		slog.Info("CreateMeme: found content-duplicate by embedding search",
-			commonconst.ACCOUNTID_LOG, request.AccountId,
+			commonconst.ACCOUNTID_LOG, accountId,
 			"id", idUuid,
 			"dupId", contentDuplicate.ImageId,
 			"ocrText", ocrTextResult,
@@ -108,14 +112,14 @@ func (a *ApiHandler) CreateMeme(
 		}
 	}
 
-	err = a.imageStorage.Save(ctx, idUuid, ocrResult.Image, ocrResult.Thumbnail.Image)
+	err = a.imageStorage.Save(ctx, idUuid, ocrResult.Image.ImageBase64, ocrResult.Thumbnail.ImageBase64)
 	if err != nil {
 		return nil, fmt.Errorf("failed to save image metadata : %w", err)
 	}
 
 	elasticMetaData := OcrResultToElastic(
 		idUuid,
-		request.AccountId,
+		accountId,
 		hash,
 		time.Now().UnixMicro(),
 		ocrResult,
@@ -140,19 +144,15 @@ func (a *ApiHandler) CreateMeme(
 
 // SearchMeme implements server.StrictServerInterface.
 func (a *ApiHandler) SearchMeme(ctx context.Context, request server.SearchMemeRequestObject) (server.SearchMemeResponseObject, error) {
-	query := request.Params.MemeQuery
+	query := request.Params.Query
+	accountId := request.AccountId
+	searchAfter := request.Params.SearchAfterSortId
+	pageSize := request.Params.PageSize
 
 	slog.Info("SearchMeme", "query", query)
-
-	matchedMetadata, err := a.metaStorage.Search(
-		ctx,
-		request.AccountId,
-		query,
-		request.Params.SearchAfterSortId,
-		request.Params.PageSize,
-	)
+	matchedMetadata, err := a.searchMeme(ctx, accountId, query, searchAfter, pageSize)
 	if err != nil {
-		return nil, fmt.Errorf("failed to search memes : %w", err)
+		return nil, fmt.Errorf("searchMeme failed: %w", err)
 	}
 
 	slog.Info("SearchMeme results",
@@ -163,32 +163,107 @@ func (a *ApiHandler) SearchMeme(ctx context.Context, request server.SearchMemeRe
 	response := make(server.SearchMeme200JSONResponse, len(matchedMetadata))
 	for index, metadataItem := range matchedMetadata {
 		slog.Debug("Search meme result item",
-			commonconst.ACCOUNTID_LOG, metadataItem.Metadata.AccountId,
-			"id", metadataItem.Metadata.ImageId,
-			"s3id", metadataItem.Metadata.S3Id,
-			"index", index,
-			"matched", render.Render(metadataItem.ResultMatched))
+			commonconst.ACCOUNTID_LOG, metadataItem.AccountId,
+			"id", metadataItem.ImageId,
+			"s3id", metadataItem.S3Id,
+			"index", index)
 
-		imageThumbUrl, err := a.imageStorage.GetUrlThumb(ctx, metadataItem.Metadata.S3Id)
+		imageThumbUrl, err := a.imageStorage.GetUrlThumb(ctx, metadataItem.S3Id)
 		if err != nil {
-			return nil, fmt.Errorf("failed to obtain thumb url for image id=%s : %w", metadataItem.Metadata.ImageId, err)
+			return nil, fmt.Errorf("failed to obtain thumb url for image id=%s : %w", metadataItem.ImageId, err)
 		}
-		imageUrl, err := a.imageStorage.GetUrl(ctx, metadataItem.Metadata.S3Id)
+		imageUrl, err := a.imageStorage.GetUrl(ctx, metadataItem.S3Id)
 		if err != nil {
-			return nil, fmt.Errorf("failed to obtain image url for image id=%s : %w", metadataItem.Metadata.ImageId, err)
+			return nil, fmt.Errorf("failed to obtain image url for image id=%s : %w", metadataItem.ImageId, err)
 		}
 
-		dto := server.SearchMemeDto{}
-		helper.ElasticToSearchMemeDto(metadataItem, &dto)
-		dto.ImageUrl = &imageUrl
-		dto.Thumbnail = new(server.SearchMemeThumb)
-		dto.Thumbnail.ThumbUrl = &imageThumbUrl
-		dto.Thumbnail.ThumbHeight = &metadataItem.Metadata.ThumbSize.Height
-		dto.Thumbnail.ThumbWidth = &metadataItem.Metadata.ThumbSize.Width
+		dto := server.SearchMemeResponseItemDto{
+			Hash: &metadataItem.Hash,
+			Id:   &metadataItem.ImageId,
+			Image: &server.ImageUrlDto{
+				Url:    imageUrl,
+				Height: metadataItem.ImageSize.Height,
+				Width:  metadataItem.ImageSize.Width,
+			},
+			Thumbnail: &server.ImageUrlDto{
+				Url:    imageThumbUrl,
+				Height: metadataItem.ThumbSize.Height,
+				Width:  metadataItem.ThumbSize.Width,
+			},
+			OcrResult: &metadataItem.Result,
+			SortId:    &metadataItem.Created,
+		}
 		response[index] = dto
 	}
 
 	return response, nil
+}
+
+func (a *ApiHandler) searchMeme(
+	ctx context.Context,
+	accountId uuid.UUID,
+	query string,
+	searchAfter *int64,
+	pageSize *int,
+) ([]*entity.ElasticImageMetaData, error) {
+	if query == "" {
+		slog.Info("SearchMeme method=ALL")
+		matchedMetadataAll, err := a.metaStorage.SearchAll(
+			ctx,
+			accountId,
+			searchAfter,
+			pageSize,
+		)
+
+		if err != nil {
+			return nil, fmt.Errorf("failed to search memes[ALL] : %w", err)
+		}
+		return matchedMetadataAll, nil
+	}
+
+	asId, err := uuid.Parse(query)
+	if err == nil {
+		slog.Info("SearchMeme method=ID")
+		matchedById, err := a.metaStorage.GetById(ctx, accountId, asId)
+		if err != nil {
+			return nil, fmt.Errorf("failed to search memes[ID] : %w", err)
+		}
+
+		if matchedById != nil {
+			return []*entity.ElasticImageMetaData{
+				matchedById,
+			}, nil
+		}
+	}
+
+	slog.Info("SearchMeme method=SIMPLE")
+	matchedMetadataSimple, err := a.metaStorage.SearchSimple(
+		ctx,
+		accountId,
+		query,
+		searchAfter,
+		pageSize,
+	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to search memes[SIMPLE] : %w", err)
+	}
+
+	if len(matchedMetadataSimple) > 0 || searchAfter != nil {
+		return matchedMetadataSimple, nil
+	}
+
+	slog.Info("SearchMeme method=FUZZY")
+	matchedMetadataFuzzy, err := a.metaStorage.SearchFuzzy(
+		ctx,
+		accountId,
+		query,
+		pageSize,
+	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to search memes[Fuzzy] : %w", err)
+	}
+
+	return matchedMetadataFuzzy, nil
 }
 
 // CheckDuplicates implements server.StrictServerInterface.
@@ -197,8 +272,14 @@ func (a *ApiHandler) CheckDuplicates(ctx context.Context, request server.CheckDu
 		a.iterateDocuments(
 			ctx,
 			request.AccountId,
-			func(ctx2 context.Context, emc *entity.ElasticMatchedContent) error {
-				a.internalCheckDuplicate(ctx2, emc.Metadata)
+			func(ctx2 context.Context, emc *entity.ElasticImageMetaData) error {
+				err := a.internalCheckDuplicate(ctx2, emc)
+				if err != nil {
+					slog.Error("internalCheckDuplicate failed",
+						"ImageId", emc.ImageId,
+						commonconst.ERR_LOG, err)
+				}
+
 				return nil
 			})
 
@@ -210,8 +291,13 @@ func (a *ApiHandler) UpdateOcr(ctx context.Context, request server.UpdateOcrRequ
 		a.iterateDocuments(
 			ctx,
 			request.AccountId,
-			func(ctx context.Context, emc *entity.ElasticMatchedContent) error {
-				a.internalUpdateOcr(ctx, emc.Metadata)
+			func(ctx context.Context, emc *entity.ElasticImageMetaData) error {
+				err := a.internalUpdateOcr(ctx, emc)
+				if err != nil {
+					slog.Error("internalUpdateOcr failed",
+						"ImageId", emc.ImageId,
+						commonconst.ERR_LOG, err)
+				}
 				return nil
 			})
 }
@@ -220,15 +306,15 @@ func (a *ApiHandler) UpdateOcr(ctx context.Context, request server.UpdateOcrRequ
 func (a *ApiHandler) GetMemeImageThumbUrl(ctx context.Context, request server.GetMemeImageThumbUrlRequestObject) (server.GetMemeImageThumbUrlResponseObject, error) {
 	memeMetadata, err := a.metaStorage.GetById(ctx, request.AccountId, request.MemeId)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("storage GetById failed: %w", err)
 	}
 
 	if memeMetadata == nil {
-		return nil, echo.ErrNotFound
+		return nil, fmt.Errorf("image for accountId=%s and MemeId=%s not found", request.AccountId, request.MemeId)
 	}
 
 	if memeMetadata.AccountId != request.AccountId {
-		return nil, echo.ErrNotFound
+		return nil, fmt.Errorf("image for accountId=%s and MemeId=%s not found", request.AccountId, request.MemeId)
 	}
 
 	url, err := a.imageStorage.GetUrlThumb(ctx, memeMetadata.S3Id)
@@ -236,8 +322,11 @@ func (a *ApiHandler) GetMemeImageThumbUrl(ctx context.Context, request server.Ge
 		return nil, err
 	}
 
-	resp := server.GetMemeImageThumbUrl200JSONResponse{}
-	resp.Url = &url
+	resp := &server.GetMemeImageThumbUrl200JSONResponse{
+		Url:    url,
+		Height: memeMetadata.ThumbSize.Height,
+		Width:  memeMetadata.ThumbSize.Width,
+	}
 	return resp, nil
 }
 
@@ -245,11 +334,15 @@ func (a *ApiHandler) GetMemeImageThumbUrl(ctx context.Context, request server.Ge
 func (a *ApiHandler) GetMemeImageUrl(ctx context.Context, request server.GetMemeImageUrlRequestObject) (server.GetMemeImageUrlResponseObject, error) {
 	memeMetadata, err := a.metaStorage.GetById(ctx, request.AccountId, request.MemeId)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("storage GetById failed: %w", err)
 	}
 
 	if memeMetadata == nil {
-		return nil, echo.ErrNotFound
+		return nil, fmt.Errorf("image for accountId=%s and MemeId=%s not found", request.AccountId, request.MemeId)
+	}
+
+	if memeMetadata.AccountId != request.AccountId {
+		return nil, fmt.Errorf("image for accountId=%s and MemeId=%s not found", request.AccountId, request.MemeId)
 	}
 
 	url, err := a.imageStorage.GetUrl(ctx, memeMetadata.S3Id)
@@ -257,8 +350,11 @@ func (a *ApiHandler) GetMemeImageUrl(ctx context.Context, request server.GetMeme
 		return nil, err
 	}
 
-	resp := server.GetMemeImageUrl200JSONResponse{}
-	resp.Url = &url
+	resp := &server.GetMemeImageUrl200JSONResponse{
+		Url:    url,
+		Height: memeMetadata.ImageSize.Height,
+		Width:  memeMetadata.ImageSize.Width,
+	}
 	return resp, nil
 }
 
@@ -266,36 +362,38 @@ func (a *ApiHandler) GetMemeImageUrl(ctx context.Context, request server.GetMeme
 func (a *ApiHandler) UpdateOcrOne(ctx context.Context, request server.UpdateOcrOneRequestObject) (server.UpdateOcrOneResponseObject, error) {
 	memeMetadata, err := a.metaStorage.GetById(ctx, request.AccountId, request.MemeId)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("storage GetById failed: %w", err)
 	}
 
 	if memeMetadata == nil {
-		return nil, echo.ErrNotFound
+		return nil, fmt.Errorf("image for accountId=%s and MemeId=%s not found", request.AccountId, request.MemeId)
 	}
 
-	a.internalUpdateOcr(ctx, memeMetadata)
+	if memeMetadata.AccountId != request.AccountId {
+		return nil, fmt.Errorf("image for accountId=%s and MemeId=%s not found", request.AccountId, request.MemeId)
+	}
+
+	err = a.internalUpdateOcr(ctx, memeMetadata)
+	if err != nil {
+		return nil, fmt.Errorf("internalUpdateOcr failed: %w", err)
+	}
+
 	return server.UpdateOcrOne200Response{}, nil
 }
 
-func (a *ApiHandler) internalCheckDuplicate(ctx context.Context, emc *entity.ElasticImageMetaData) {
+func (a *ApiHandler) internalCheckDuplicate(ctx context.Context, emc *entity.ElasticImageMetaData) error {
 	id := emc.ImageId
 	embedding := emc.EmbeddingV1
 	slog.Info("Check-duplicate",
 		"id", id.String())
 
 	if embedding == nil {
-		slog.Info("Check-duplicate: NO EMBEDDING:",
-			"id", id.String())
-		return
+		return fmt.Errorf("embedding is nil. id=%s", id)
 	}
 
-	embeddingFoundImage, err := a.metaStorage.GetByEmbeddingV1All(ctx, embedding, 10)
+	embeddingFoundImage, err := a.metaStorage.GetByEmbeddingV1(ctx, emc.AccountId, embedding, 10)
 	if err != nil {
-		slog.Error("Check-duplicate: failed to search image embedding duplicates ",
-			"id", id.String(),
-			commonconst.ERR_LOG, err)
-
-		return
+		return fmt.Errorf("GetByEmbeddingV1All failed: %w", err)
 	}
 
 	for i, item := range embeddingFoundImage {
@@ -310,10 +408,11 @@ func (a *ApiHandler) internalCheckDuplicate(ctx context.Context, emc *entity.Ela
 				"id", item.ImageId)
 		}
 	}
+
+	return nil
 }
 
-func (a *ApiHandler) internalUpdateOcr(ctx context.Context, emc *entity.ElasticImageMetaData) {
-
+func (a *ApiHandler) internalUpdateOcr(ctx context.Context, emc *entity.ElasticImageMetaData) error {
 	id := emc.ImageId
 	accountId := emc.AccountId
 	hash := emc.Hash
@@ -323,66 +422,78 @@ func (a *ApiHandler) internalUpdateOcr(ctx context.Context, emc *entity.ElasticI
 	slog.Info("UpdateOcr: checking image",
 		"id", id)
 
-	img, err := a.imageStorage.GetImage(ctx, s3id)
+	base64ImgData, err := a.imageStorage.GetImage(ctx, s3id)
 	if err != nil {
-		slog.Info("Failed to read image from storage",
-			"id", id,
-			commonconst.ERR_LOG, err)
-
-		return
+		return fmt.Errorf("storage GetImage failed: %w", err)
 	}
 
-	ocrResult, err := a.ocr.DoOcr(ctx, id, img)
+	item := &entity.Image{
+		ImageBase64: *base64ImgData,
+		Width:       emc.ImageSize.Width,
+		Height:      emc.ImageSize.Height,
+	}
+
+	ocrResult, err := a.ocr.DoOcr(ctx, item)
 	if err != nil {
-		slog.Info("Failed to ocr image",
-			"id", id,
-			commonconst.ERR_LOG, err)
-		return
+		return fmt.Errorf("doOcr failed: %w", err)
 	}
 
 	elasticObject := OcrResultToElastic(id, accountId, hash, created, ocrResult)
 	err = a.metaStorage.Save(ctx, elasticObject)
 	if err != nil {
-		slog.Info("Failed to save new image metadata",
-			"id", id,
-			commonconst.ERR_LOG, err)
-		return
+		return fmt.Errorf("doOcr failed: %w", err)
 	}
+
+	return nil
 }
 
-func (a *ApiHandler) iterateDocuments(ctx context.Context, accountId uuid.UUID, callback func(context.Context, *entity.ElasticMatchedContent) error) error {
-	items, err := a.metaStorage.Search(ctx, accountId, "", nil, addr(1000))
-	for err == nil && len(items) > 0 {
+func (a *ApiHandler) iterateDocuments(
+	ctx context.Context,
+	accountId uuid.UUID,
+	callback func(context.Context, *entity.ElasticImageMetaData) error,
+) error {
+	items, err := a.metaStorage.SearchAll(ctx, accountId, nil, commonhelper.Addr(1000))
+	if err != nil {
+		return fmt.Errorf("search failed: %w", err)
+	}
+
+	for len(items) > 0 {
 		for _, item := range items {
 			err = callback(ctx, item)
+			if err != nil {
+				return fmt.Errorf("iterateDocuments callback failed: %w", err)
+			}
 		}
+
 		if len(items) > 0 {
-			items, err = a.metaStorage.Search(ctx, accountId, "", &items[len(items)-1].Metadata.Created, addr(1000))
+			items, err = a.metaStorage.SearchAll(ctx, accountId, &items[len(items)-1].Created, commonhelper.Addr(1000))
+			if err != nil {
+				return fmt.Errorf("search failed: %w", err)
+			}
 		}
 	}
 
-	if err != nil {
-		return err
-	}
 	return nil
 }
 
 func (a *ApiHandler) findHashDuplicates(
 	ctx context.Context,
+	accountId uuid.UUID,
 	hash string,
-) (*entity.ElasticImageMetaData, error) {
-	return a.metaStorage.GetByHashAll(ctx, hash)
+) ([]*entity.ElasticImageMetaData, error) {
+	return a.metaStorage.GetByHash(ctx, accountId, hash, nil)
 }
 
 func (a *ApiHandler) findContentDuplicates(
 	ctx context.Context,
+	accountId uuid.UUID,
 	ocrResult *OcrProcessedResult,
 ) (*entity.ElasticImageMetaData, error) {
 	embedding := &entity.ElasticEmbeddingV1{
 		Data:  &ocrResult.Embedding.Data,
 		Model: ocrResult.Embedding.Model,
 	}
-	results, err := a.metaStorage.GetByEmbeddingV1All(ctx, embedding, 1)
+	results, err := a.metaStorage.GetByEmbeddingV1(ctx, accountId, embedding, 1)
 	if err != nil {
 		return nil, err
 	}
@@ -423,6 +534,10 @@ func OcrResultToElastic(idUuid uuid.UUID, accountId uuid.UUID, hash string, crea
 		ImageId:   idUuid,
 		S3Id:      idUuid,
 		AccountId: accountId,
+		ImageSize: &entity.ElasticSizes{
+			Height: ocrResult.Image.Height,
+			Width:  ocrResult.Image.Width,
+		},
 		ThumbSize: &entity.ElasticSizes{
 			Height: ocrResult.Thumbnail.Height,
 			Width:  ocrResult.Thumbnail.Width,
